refactored admin, added search
