+++ src/main/resources/templates/about.html
@@ -24,6 +24,7 @@
          <li>Spring Boot</li>
          <li>Thymeleaf</li>
          <li>Bootstrap</li>
+         <li>JavaScript</li>
          <li>HTML</li>
          <li>CSS</li>
       </ul>
