+++ src/main/resources/templates/features.html
@@ -14,6 +14,7 @@
       <ul>
          <li>Browse recipes (list recipe cards w/ link to recipe)</li>
          <li>Search recipes (return recipe cards)</li>
+         <li>Advanced Search (return recipe cards)</li>
          <li>View full recipe (display recipe)</li>
       </ul>
       <h5>Administration</h5>
@@ -37,7 +38,7 @@
          </li>
       </ul>
 
-      <p>Extensive error checking for admin included for adding, editing and deleting. Including modal popup warnings and error/success messages on actions.</p>
+      <p>Extensive error checking for admin included for adding, editing and deleting. Including modal popup warnings and error/success messages.</p>
    </div>
 
    <footer th:replace="fragments :: footer"></footer>
