+++ src/main/resources/templates/fragments.html
@@ -23,10 +23,10 @@
                         <a class="nav-link px-3" th:href="@{/}">Home</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link px-3" th:href="@{/recipes/}">Recipes</a>
+                        <a class="nav-link px-3" th:href="@{/recipes}">Recipes</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link px-3" th:href="@{/search/}">Search</a>
+                        <a class="nav-link px-3" th:href="@{/search}">Search</a>
                     </li>
                     <li class="nav-item">
                         <a class="nav-link px-3" th:href="@{/about}">About</a>
@@ -55,14 +55,15 @@
                 <div class="text-white">
                     <h1 class="mb-3"><a class="text-white text-reset text-decoration-none" th:href="@{/}">THE PANTRY</a></h1>
                     <h4 class="mb-3">Find recipe inspiration based on the food in your pantry.</h4>
-                    <form method="post" action="search">
+                    <form method="post" th:action="@{/recipes/results}">
                         <div class="w-50 mx-auto">
                             <div class="input-group">
-                                <input class="form-control" name="search" placeholder="Let's raid the Pantry!" />
+                                <input class="form-control" name="searchTerm" placeholder="Let's raid the Pantry!" />
                                 <input class="btn btn-outline-light" type="submit" value="Search" />
                             </div>
                         </div>
                     </form>
+                    <div class="pt-2"><a th:href="@{/search}" class="text-muted text-reset text-decoration-none">Advanced Search</a></div>
                 </div>
             </div>
         </div>
@@ -75,8 +76,8 @@
             </p>
             <ul class="nav col-md-6 justify-content-end">
                 <li class="nav-item"><a class="nav-link px-2 text-muted" th:href="@{/}">Home</a>
-                <li class="nav-item"><a class="nav-link px-2 text-muted" th:href="@{/list/}">List</a></li>
-                <li class="nav-item"><a class="nav-link px-2 text-muted" th:href="@{/search/}">Search</a></li>
+                <li class="nav-item"><a class="nav-link px-2 text-muted" th:href="@{/recipes}">Recipes</a></li>
+                <li class="nav-item"><a class="nav-link px-2 text-muted" th:href="@{/search}">Search</a></li>
                 <li class="nav-item"><a class="nav-link px-2 text-muted" th:href="@{/about}">About</a></li>
                 <li class="nav-item"><a class="nav-link px-2 text-muted" th:href="@{/features}">Features</a></li>
             </ul>
