+++ src/main/resources/templates/index.html
@@ -12,7 +12,7 @@
         <h1>Cooking from the Pantry</h1>
         <p>Browse our recipes for just the right meal.</p>
 
-        <div class="row row-cols-4 g-4">
+        <div class="row row-cols-4 g-4 pt-1">
             <div class="col" th:each="recipe : ${recipes}">
                 <div class="card border shadow bg-white rounded h-100">
                     <a th:href="@{'/recipe/' + ${recipe.id}}"><img th:src="@{'/images/' + ${recipe.image}}" class="card-img-top"></a>
