+++ src/main/resources/templates/recipes.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="https://www.thymeleaf.org/">
+<head th:replace="fragments :: head"></head>
+<body>
+
+<header>
+    <div th:replace="fragments :: nav"></div>
+    <div th:replace="fragments :: banner"></div>
+</header>
+
+<div class="container body-content py-3">
+    <h1 th:text="${title}"></h1>
+
+    <p th:if="${message}" th:text="${message}" th:class="${class}"></p>
+
+    <div class="row row-cols-4 g-4 pt-1">
+        <div class="col" th:each="recipe : ${recipes}">
+            <div class="card border shadow bg-white rounded h-100">
+                <a th:href="@{'/recipe/' + ${recipe.id}}"><img th:src="@{'/images/' + ${recipe.image}}" class="card-img-top"></a>
+                <div class="card-body">
+                    <h5 class="card-title"><a th:href="@{'/recipe/' + ${recipe.id}}" th:text="${recipe.name}" class="text-reset text-decoration-none fw-bolder"></a></h5>
+                    <p class="card-text" th:text="${#strings.abbreviate(recipe.description,85)}"></p>
+                </div>
+            </div>
+        </div>
+    </div>
+
+</div>
+
+<footer th:replace="fragments :: footer"></footer>
+
+</body>
+</html>
